Changes to support arm64:
Here are the drivers that I've tested to be working.

1. BCM2709 DMA API - Needed to make the FB video driver work correctly without any changes.

   This driver contains two dsb instructions where are not supported in the 64bit build environment.
   Convert these to the more portable wmb and rmb macros.

2. DWC_OTG USB

   Driver uses FIQ which is not support in the box on arm64.  Since the CPUs are much faster now then then pi 1,
   on arm64 only this was converted to a standard irq.  The irq drivers were modifiied to round robin dispatch irqs
   between cores 0 and 1 for more concurrency on arm64 only.

   Even though FIQ is not used on 64bit, the driver still outproforms the upstream driver.

3. FT5406 touchscreen support

   Driver was reading a writing a single byte register which was cause a exception on 64 bit.  Change these to essentially a
   memcpy call with a small loss of performance.

4. Limited VCHIQ support.

   Driver was storing pointers in the shared memory between the gpu and arm cores.  On 64 bit only, change these to offsets.

5. HDMI audio.

   Driver was storing a callback in shared memory.  Luckely a spare field was not being used.  On 64bit only, split the callback
   address across two uint32_t fields.

These drivers just work without any needed changes
1. bcm2835-mmc
2. bcm2835-sdhost
3. Builtin bluetooth
4. Builtin wifi
5. VC4 DRM driver for advanced video and opengl support.

+++ include/linux/platform_data/dma-bcm2708.h
@@ -63,13 +63,13 @@
 #define BCM_DMA_FEATURE_COUNT		4
 
 struct bcm2708_dma_cb {
-	unsigned long info;
-	unsigned long src;
-	unsigned long dst;
-	unsigned long length;
-	unsigned long stride;
-	unsigned long next;
-	unsigned long pad[2];
+	uint32_t info;
+	uint32_t src;
+	uint32_t dst;
+	uint32_t length;
+	uint32_t stride;
+	uint32_t next;
+	uint32_t pad[2];
 };
 
 struct scatterlist;
